tikfig/requirements.txt
Flask==3.0.0
Jinja2==3.1.2
PyYAML==6.0.1
requests==2.31.0
